+++ src/Resources/Components/ExerciseDropdown/ExerciseDropdownStyle.js
@@ -4,7 +4,6 @@ const styles = StyleSheet.create({
   container: {
     borderWidth: 1,
     borderRadius: 50,
-    borderColor: "#ccc",
     marginBottom: 16,
     width: '80%',
   },